==> notebooks/1. Data preprocessing.ipynb <==
import os
import numpy as np
import pandas as pd

# Suppress warnings
import warnings
warnings.filterwarnings("ignore")

RAW_DATA_DIR = '../data/raw'
INTERIM_DIR = '../data/interim'

### Initial processing of dataset

title_akas = pd.read_csv('../data/raw/title_akas.tsv', sep='\t')
title_basics = pd.read_csv('../data/raw/title_basics.tsv', sep='\t')
title_crew = pd.read_csv('../data/raw/title_crew.tsv', sep='\t')
title_principals = pd.read_csv('../data/raw/title_principals.tsv', sep='\t')
title_ratings = pd.read_csv('../data/raw/title_ratings.tsv', sep='\t')
name_basics = pd.read_csv('../data/raw/name_basics.tsv', sep='\t')

print("Number of titles    : {}".format(len(title_basics)))
print("Number of ratings   : {}".format(len(title_ratings)))
print("Number of actors    : {}".format(len(name_basics)))
print("Number of directors : {}".format(len(title_crew['directors'].unique())))
print("Number of writers   : {}".format(len(title_crew['writers'].unique())))

# Listing all possible Types (we only need movie)

title_basics['titleType'].unique()

#### Keep the movie titles that have a rating

# Keep only the titles that have a rating

titles = title_basics.merge(title_ratings, on='tconst')
titles

## Some Processes

- Keep 'movie' type only
- Convert the 'startYear' to integer values
- Drop the 'endYear' field as it is irrelevant for movies

def process_null_values(df):
    '''
    Change non standard Null value('\\N') to np.NaN so that pandas recognizes it
    
    params:
        - df (pd.DataFrame): Dataframe to process
    
    returns:
        - pd.DataFrame: Dataframe with np.NaN in place of Null
    '''
    return df.applymap(lambda x: np.NaN if (x=='\\N') else x)

# Taking care of NaN values
titles = process_null_values(titles)

# Drop 'endYear'
del titles['endYear']

# Filtering movies types
movies = titles[titles['titleType'].isin(['movie'])]

# Processing null values in startYear (average year)
mean_startYear = int(movies['startYear'].dropna().astype(int).mean())

movies['startYear'] = movies['startYear'].apply(lambda x: int(x) if (not pd.isna(x)) else mean_startYear)

# Resetting row numbers
movies.reset_index(drop=True, inplace=True)
movies

#### Overview of columns having null values

movies.isnull().sum()

#### Drop missing values for genres

# Drop data which has Null value in genres
movies = movies.loc[movies['genres'].notnull()]

#### Replacing Null values in runtimeMinutes by mean value

mean_runtime = int(movies['runtimeMinutes'].dropna().astype(int).mean())

movies['runtimeMinutes'] = movies['runtimeMinutes'].apply(lambda x: mean_runtime if (pd.isna(x)) else x)

movies.to_csv(os.path.join(INTERIM_DIR, 'movies.csv'))

### Processing title_crew, title_principals files

title_crew = process_null_values(title_crew)
title_crew

title_principals = process_null_values(title_principals)
title_principals

#### Save only relevant data

# Filtering the crew and principals data from only relevant titles

# Add the titles we are considering to a set and then filter the dataset
considered_titles = set()
considered_titles.update(movies['tconst'].values)

title_principals = title_principals[title_principals['tconst'].isin(considered_titles)]
title_crew = title_crew[title_crew['tconst'].isin(considered_titles)]

print("Length of processed title_crew: {}".format(len(title_crew)))
print("Length of processed title_principals: {}".format(len(title_principals)))

# Saving files as csvs to interim data folder

title_crew.to_csv(os.path.join(INTERIM_DIR, 'title.crew.csv'), index=False)
title_principals.to_csv(os.path.join(INTERIM_DIR, 'title.principals.csv'), index=False)

#### Filtering the names_basics dataframe for including only the names associated with considered movies for faster processing

**NOTE**: This filtering is only for saving time in prototyping further in the pipeline as I plan to calculate features for each person. In production, I would retain the entire name list and calculate the features on them as well in case a movie has a crew member who hasnt worked in a movie before (eg. an actor who only did TV series up until now).

**Observation:** IMDB lists the top billed actors ('star cast' or protagonist) first and the list decreases in order of importance

**Design decision:** From the cast of the movie, for genres computation, we only consider the **top 5 names** ('star cast' or actors playing most important characters) to reduce our calculations significantly. We already processed _title_principles_ to allow the same

names = process_null_values(name_basics)
names

considered_name_ids = set()

def split_nameIDs(nconst_str):
    '''
    Split the nconst string by commas and filter out empty strings
    
    params:
        - nconst_str (str): String of comma separated nameIDs
    
    returns:
        - (list): List of nameIDs
    '''
    return [nameID for nameID in nconst_str.split(',') if nameID!='']

# Adding all directors
title_crew['directors'].dropna().apply(lambda x: considered_name_ids.update(split_nameIDs(x)))

# Adding all writers
title_crew['writers'].dropna().apply(lambda x: considered_name_ids.update(split_nameIDs(x)))

# Adding principal crew members
considered_name_ids.update(title_principals['nconst'].dropna().values)

print(len(considered_name_ids))

names = names[names['nconst'].isin(considered_name_ids)]
names.reset_index(drop=True, inplace=True)
names

names.to_csv(os.path.join(INTERIM_DIR, 'names.basics.csv'))

#### We need all the ratings available because the plan is to calculate features for each person using the titles they are known for

title_ratings = process_null_values(title_ratings)
title_ratings

title_ratings.isnull().sum()

title_ratings.to_csv(os.path.join(INTERIM_DIR, 'title_ratings.csv'))

title_akas = process_null_values(title_akas)
title_akas

title_akas.isnull().sum()

# Filter the dataframe to include titles we consider and no more
title_akas = title_akas[title_akas['titleId'].isin(considered_titles)][['titleId', 'ordering']]
title_akas

title_akas.to_csv(os.path.join(INTERIM_DIR, 'title.akas.csv'))
