==> notebooks/2. Feature Engineering.ipynb <==
import os
import numpy as np
import pandas as pd
from math import ceil
from tqdm import tqdm
from joblib import Parallel, delayed
from sklearn.preprocessing import MultiLabelBinarizer

# Suppress warnings
import warnings
warnings.filterwarnings("ignore")

INTERIM_DIR = '..\\data\\interim'
FINAL_DIR = '..\\data\\final'

#### Reading all data files

movies_df = pd.read_csv(os.path.join(INTERIM_DIR, 'movies.csv'))
title_ratings_df = pd.read_csv(os.path.join(INTERIM_DIR, 'title_ratings.csv'))
title_akas_df = pd.read_csv(os.path.join(INTERIM_DIR, 'title.akas.csv'))
title_principals_df = pd.read_csv(os.path.join(INTERIM_DIR, 'title.principals.csv'))
title_crew_df = pd.read_csv(os.path.join(INTERIM_DIR, 'title.crew.csv'))
names_df = pd.read_csv(os.path.join(INTERIM_DIR, 'names.basics.csv'))

### New Feature 1: Popularity Index of Actors/Directors/Crew

#### Popularity Index of a Person

**Description:**  
  
  1. For each person in the names_df, we have a list of titleID.
  2. Using the list of titleIDs, we can get a list of each person's ratings from the title_ratings_df.
  
**Assumption:**
  1. The main assumption is that the ratings associated with the title are related with how popular the movie is.
  2. We are going to infer the popularity of the person from the popularity of the titles, which assumes that the titles represent their popularity entirely.

**Intuition:**
  
  1. **Mean** of the list of ratings gives us an idea of how popular the person is. Even if the person generally is average, but they have done one/two movies that perform exceptionally well (according to ratings), the mean will capture that.
  2. **Median** of the list of ratings captures the general consensus of the person's performance that is how they are usually expected to perform.
  3. **Standard Deviation** of these numbers will characterize the distribution of the ratings. For Eg. if the mean of the ratings is high, and the standard deviation is low, it means that the person consistently has performed exceptionally well. Similarly if the mean is low but standard deviation is high, it means that the person has been 'average' in their performance sometimes.

names_df

def calculate_pop_ind_parallel(names_df, title_ratings_df):
    '''
    Calculate the popularity index of a person using titles they 
    are known for in parallel.
    
    params:
        - names_df         (DataFrame): names table
        - title_ratings_df (DataFrame): ratings table
    
    returns:
        - output      (list of tuples): list of popularity indices
    '''
    name_IDs = names_df['nconst'].tolist()
    
    # Get the number of logical processors and distribute the load
    num_cores = os.cpu_count()
    names_per_core = ceil(len(name_IDs)/num_cores)
    print("names_per_core: {}".format(names_per_core))
    
    name_list_per_core = [name_IDs[i*names_per_core: (i+1)*names_per_core]
                          for i in range(num_cores)]
    
    # Execute in parallel
    output_list = Parallel(n_jobs=-1, verbose=30, max_nbytes=None)(delayed(get_pop_ind)(nameID_list, 
                                                                                        names_df.copy(), 
                                                                                        title_ratings_df.copy())
                                                                  for nameID_list in name_list_per_core)
    
    # Aggregate results from all parallel workers
    result = []
    [result.extend(list_) for list_ in output_list if list_!=[]]
    
    return result

def get_pop_ind(nameID_list, names_df, title_ratings_df):
    '''
    Calculate the popularity index values from the names_df and title ratings 
    
    Steps (for each nameID):
    1. Collect the titleIDs that a person is known for 
    2. Collect the ratings of these titles from the ratings table
    3. Calculate the mean and std. deviation for those ratings and return
    
    params:
        - nameID_list             (list): list of nameIDs
        - names_df           (DataFrame): names table
        - title_ratings_df   (DataFrame): ratings table
    
    returns:
        - pop_inds      (list of tuples): list of popularity indices
    '''
    
    pop_inds = []
    
    for nameID in nameID_list:
        
        pop_ind_mean, pop_ind_med, pop_ind_std = 0,0,0
        
        # get the record for nameID from names_df
        knownfor = names_df[names_df['nconst']==nameID]['knownForTitles'].iloc[0]
        
        if not pd.isna(knownfor):
            
            # get the titles that the person is known for
            titles = [title for title in knownfor.split(',') if title!='']
            # get the ratings for those titles
            ratings = title_ratings_df[title_ratings_df['tconst'].isin(titles)]['averageRating']
            
            pop_ind_mean = ratings.mean()
            pop_ind_std = ratings.std()
            pop_ind_med = ratings.median()
        
        pop_inds.append((pop_ind_mean, pop_ind_med, pop_ind_std))
    
    return pop_inds

popularity_index = calculate_pop_ind_parallel(names_df, title_ratings_df)

popularity_index[0:10]

mean, median, std = zip(*popularity_index)

print(len(mean), len(median), len(std))

#### Setting the values in the Dataframe

names_df['pop_ind_mean'] = list(mean)
names_df['pop_ind_median'] = list(median)
names_df['pop_ind_std'] = list(std)

names_df[['pop_ind_mean', 'pop_ind_median','pop_ind_std']]

#### Let us test how these values behave for extremely well known actors

stars = ['Tom Cruise', 'Leonardo DiCaprio', 'Angelina Jolie', 'Nicole Kidman', 'Brad Pitt', 'Johnny Depp']

names_df[names_df['primaryName'].isin(stars)][['primaryName','pop_ind_mean', 'pop_ind_median','pop_ind_std']]

#### As expected, the values of the popularity index are fairly high for these actors and the standard deviation is low. This suggests consistently great performance as we had imagined.

names_df.to_csv(os.path.join(FINAL_DIR, 'names.basics.csv'), index=False)

### Infering the popularity of the movie from the popularity of its cast and crew

**Description:**  
1. We essentially want to estimate the popularity of a movie based on the popularity of its actors/directors/writers
2. For this, we analyze the means and standard deviations of each of the values that make up our popularity index for persons
3. These 6 values collectively will represent the popularity index of the movie.
4. By calculating the mean and standard deviation, we effectively summarize the popularity indices of its cast

**NOTE:** This again takes into account the crew and only the top 5 actors as we have filtered the title.principals table

def split_nameIDs(nconst_str):
    '''
    Split the nconst string by commas and filter out empty strings
    
    params:
        - nconst_str (str): String of comma separated nameIDs
    
    returns:
        - (list): List of nameIDs
    '''
    return [nameID for nameID in nconst_str.split(',') if nameID!='']

def get_movie_popind(titleID_list, 
                     title_principals=title_principals_df, 
                     title_crew=title_crew_df, 
                     names_df=names_df):
    '''
    Get a feature vector representing the popularity index of movie based
    on the popularity indices of its cast and crew 
    
    params:
        - titleID_list            (list): list of titleIDs
        - title_principals   (DataFrame): movie actors table
        - title_crew         (DataFrame): title crew data
        - names_df           (DataFrame): names table
    
    returns:
        - pop_indexes   (list of tuples): popularity indices of movies
    '''
    pop_indexes = []
    
    for titleID in titleID_list:
    
        pi_mean_mean, pi_mean_std = 0,0
        pi_med_mean, pi_med_std = 0,0
        pi_std_mean, pi_std_std = 0,0

        people_involved = set()

        # Adding top 5 actors associated with the movie to the set
        people_involved.update(title_principals[title_principals['tconst']==titleID]['nconst'].values)

        # Adding writers and directors
        crew = title_crew[title_crew['tconst']==titleID]

        if not pd.isna(crew['writers'].iloc[0]):
            people_involved.update(split_nameIDs(crew['writers'].iloc[0]))

        if not pd.isna(crew['directors'].iloc[0]):
            people_involved.update(split_nameIDs(crew['directors'].iloc[0]))

        # Choose the people involved in the names list
        people_involved_df = names_df[names_df['nconst'].isin(people_involved)]

        # Calculate the statistics of the people's popularity indexes
        pi_mean = people_involved_df['pop_ind_mean']
        pi_med = people_involved_df['pop_ind_median']
        pi_std = people_involved_df['pop_ind_std']

        pi_mean_mean, pi_mean_std = pi_mean.mean(), pi_mean.std()
        pi_med_mean, pi_med_std = pi_med.mean(), pi_med.std()
        pi_std_mean, pi_std_std = pi_std.mean(), pi_std.std()
        
        pop_indexes.append((pi_mean_mean, pi_mean_std,\
                           pi_med_mean, pi_med_std,\
                           pi_std_mean, pi_std_std))
    
    return pop_indexes

def calculate__movie_pop_ind_parallel(movies_df,
                                      title_principals=title_principals_df, 
                                      title_crew=title_crew_df, 
                                      names_df=names_df):
    '''
    Calculate the popularity index of a movie in parallel
    
    params:
        - movies_df          (DataFrame): dataset
        - title_principals   (DataFrame): movie actors table
        - title_crew         (DataFrame): title crew data
        - names_df           (DataFrame): names table
    
    returns:
        - result        (list of tuples): popularity indices of movies
    '''
    title_IDs = movies_df['tconst'].tolist()
    
    # Distribute load across CPU cores
    num_cores = os.cpu_count()
    titles_per_core = ceil(len(title_IDs)/num_cores)
    print("titles_per_core: {}".format(titles_per_core))
    
    titles_list_per_core = [title_IDs[i*titles_per_core: (i+1)*titles_per_core]
                          for i in range(num_cores)]
    
    # Execute in parallel
    output_list = Parallel(n_jobs=-1, verbose=30, max_nbytes=None)(delayed(get_movie_popind)(titleID_list,
                                                                                        title_principals_df.copy(),
                                                                                        title_crew_df.copy(),
                                                                                        names_df.copy()) 
                                                                  for titleID_list in titles_list_per_core)
    
    # Aggregate results
    result = []
    [result.extend(list_) for list_ in output_list if list_!=[]]
    
    return result

movie_pop_ind = calculate__movie_pop_ind_parallel(movies_df)

pi_mean_mean, pi_mean_std,\
pi_med_mean, pi_med_std,\
pi_std_mean, pi_std_std = zip(*movie_pop_ind)

variables = [pi_mean_mean, pi_mean_std,\
             pi_med_mean, pi_med_std,\
             pi_std_mean, pi_std_std]

names = ['pi_mean_mean', 'pi_mean_std',\
         'pi_med_mean', 'pi_med_std',\
         'pi_std_mean', 'pi_std_std']

pi_dict = list(zip(names, variables))

for (column_name, values) in pi_dict:
    
    movies_df[column_name] = values

movies_df.head()

na_values = dict(zip(names, [0]*len(names)))
print(na_values)
movies_df = movies_df.fillna(value=na_values)
movies_df.isnull().sum()

movies_df.to_csv(os.path.join(FINAL_DIR, 'movies.csv'), index=False)

### Feature 2: One hot encoding of genres

**Description:**

   1. The _genres_ column has at most 3 genres associated with the movie.
   2. We transform this column into a sparse one-hot encoded vector _(since there are 28 unique genres)_ te represent each movie.
    
**Intuition:**

   1. Each vector will represent the genre combination of the movie and might turn out to be distinguishing if the model is able to associate the cast and the genres to predict the movie's average rating


# List all genres

genres = pd.read_csv(os.path.join(INTERIM_DIR, 'genres.csv'))['genres'].values
genres

binarizer = MultiLabelBinarizer()
binarizer.fit([list(genres)])
binarizer.classes_

mllabels = binarizer.transform(movies_df['genres'].apply(lambda x: [x for x in x.split(',') if x!=[]]))

labels = pd.DataFrame(data=mllabels, columns=binarizer.classes_)

pd.concat([movies_df, labels], axis=1)

movies_df = pd.concat([movies_df, labels], axis=1)

movies_df.to_csv(os.path.join(FINAL_DIR, 'movies.csv'), index=False)
